+++ NEWS.md
@@ -13,6 +13,7 @@
 
   * Better error-checking in multinomial parameters definition.
   * Handle cases when parameters depend on `state_time` but thoses parameters are not used by a strategy.
+  * Empty calls to `modify()` are supported.
 
 ### Backend changes